pyproject.toml
[build-system]
requires = [
    "setuptools",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "airflow-provider-mailtrap"
authors = [
    { name = "Mailtrap", email = "support@mailtrap.io" },
]
license = { text = "Apache License 2.0" }
description = "Apache Airflow provider for Mailtrap.io email service"
classifiers = [
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dynamic = [
    "version"
]
requires-python = ">=3.9"
dependencies = [
    "apache-airflow>=2.4",
    "mailtrap",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-mock",
    "requests-mock",
    "ruff",
]

[project.urls]
Homepage = "https://mailtrap.io"
Source = "https://github.com/railsware/airflow-provider-mailtrap"
Documentation = "https://github.com/railsware/airflow-provider-mailtrap#readme"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.entry-points.apache_airflow_provider]
provider_info = "mailtrap_provider.__init__:get_provider_info"

[tool.setuptools]
include-package-data = false

[tool.setuptools.dynamic]
version = { attr = "mailtrap_provider.__version__" }

[tool.setuptools.packages.find]
exclude = [
    "*tests.*",
    "*tests"
]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
